+++ packages/web-core/src/entities/packets/inworld_packet.entity.ts
@@ -1,5 +1,6 @@
 import {
   Agent,
+  ControlEventAction,
   DataChunkDataType,
   InworldPacket as ProtoPacket,
   LoadCharactersCharacterName,
@@ -40,9 +41,9 @@ export interface SceneMutation {
   name?: string;
   description?: string;
   displayName?: string;
-  characterNames?: string[];
+  addedCharacterNames?: string[];
+  removedCharacterIds?: string[];
   loadedCharacters?: Character[];
-  addedCharacters?: Character[];
 }
 
 export class InworldPacket {
@@ -224,26 +225,22 @@ export class InworldPacket {
             name: proto.mutation.loadScene.name,
           }),
           ...(proto.mutation?.loadCharacters && {
-            characterNames: proto.mutation.loadCharacters.name.map(
+            addedCharacterNames: proto.mutation.loadCharacters.name.map(
               (c: LoadCharactersCharacterName) => c.name,
             ),
           }),
-          ...(proto.sessionControlResponse?.loadedScene && {
-            loadedCharacters:
-              proto.sessionControlResponse.loadedScene.agents.map(
-                (agent: Agent) => Character.fromProto(agent),
-              ),
-            name: proto.sessionControlResponse.loadedScene.sceneName,
-            description:
-              proto.sessionControlResponse.loadedScene.sceneDescription,
-            displayName:
-              proto.sessionControlResponse.loadedScene.sceneDisplayName,
+          ...(proto.mutation?.unloadCharacters && {
+            removedCharacterIds: proto.mutation.unloadCharacters.agents.map(
+              (c: Agent) => c.agentId,
+            ),
           }),
-          ...(proto.sessionControlResponse?.loadedCharacters && {
-            addedCharacters:
-              proto.sessionControlResponse.loadedCharacters.agents.map(
-                (agent: Agent) => Character.fromProto(agent),
-              ),
+          ...(proto.control?.currentSceneStatus && {
+            name: proto.control.currentSceneStatus.sceneName,
+            description: proto.control.currentSceneStatus.sceneDescription,
+            displayName: proto.control.currentSceneStatus.sceneDisplayName,
+            loadedCharacters: proto.control.currentSceneStatus.agents.map(
+              (agent: Agent) => Character.fromProto(agent),
+            ),
           }),
         },
       }),
@@ -251,7 +248,17 @@ export class InworldPacket {
   }
 
   private static getType(packet: ProtoPacket) {
-    if (packet.text) {
+    if (
+      packet.mutation?.loadScene ||
+      packet.mutation?.loadCharacters ||
+      packet.mutation?.unloadCharacters
+    ) {
+      return InworldPacketType.SCENE_MUTATION_REQUEST;
+    } else if (
+      packet.control?.action === ControlEventAction.CURRENT_SCENE_STATUS
+    ) {
+      return InworldPacketType.SCENE_MUTATION_RESPONSE;
+    } else if (packet.text) {
       return InworldPacketType.TEXT;
     } else if (
       packet.dataChunk &&
@@ -273,13 +280,6 @@ export class InworldPacket {
       return InworldPacketType.CANCEL_RESPONSE;
     } else if (packet.action?.narratedAction) {
       return InworldPacketType.NARRATED_ACTION;
-    } else if (
-      packet.sessionControlResponse?.loadedScene ||
-      packet.sessionControlResponse?.loadedCharacters
-    ) {
-      return InworldPacketType.SCENE_MUTATION_RESPONSE;
-    } else if (packet.mutation?.loadScene || packet.mutation?.loadCharacters) {
-      return InworldPacketType.SCENE_MUTATION_REQUEST;
     } else {
       return InworldPacketType.UNKNOWN;
     }